update project structures_typedef

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -25,7 +25,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	}
 
 
-	for (j = 0; j < n; j++)
+	for (j = 0; j < (int)n; j++)
 	{
 		str[i + j] = s2[j];
 	}
